requirements.txt
# Core Dependencies
streamlit>=1.31.0
python-dotenv>=1.0.0

# LangChain Components
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10

# Vector Storage
faiss-cpu>=1.7.4

# PDF Processing
PyPDF2>=3.0.0

# OpenAI
openai>=1.12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Environment Management
python-dotenv>=1.0.0